requirements.txt
streamlit==1.32.0
pandas==2.2.0
pyarrow==15.0.0
pyyaml==6.0.1
numpy==1.26.4
matplotlib==3.8.3